refactor(floot): extract shared buffered Far reader for stream wires
The reply, transcript, and audio wires each hand-rolled the same buffered
Far StreamReader (buffer + wake/park loop + terminal detection +
return/throw consumer-stop handling), differing only in event vocabulary.

Extract that mechanism into src/buffered-channel.js (makeBufferedReader)
and reduce each site to a thin domain writer over the shared push. The
two onClose shapes are preserved: TTS passes it at construction, STT wires
it via setOnClose, and the reply channel omits it.
